+++ app/src/v/deleteShelter.ts
@@ -23,9 +23,11 @@ fillSelectWithOptions(shelterSelection, ShelterStorage.instances, {
 deleteButton.addEventListener("click", () => {
   const id = shelterSelection.value;
   if (id) {
-    ShelterStorage.destroy(id);
+    if (confirm("Do you really want to delete this Movie?")) {
+      ShelterStorage.destroy(id);
 
-    // remove deleted shelter from selection
-    shelterSelection.remove(shelterSelection.selectedIndex);
+      // remove deleted shelter from selection
+      shelterSelection.remove(shelterSelection.selectedIndex);
+    }
   }
 });
